LayoutTests/http/tests/security/mixedcontent-geolocation-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/resources/geolocation-over-mixed-content.html requested insecure content from http://127.0.0.1:8080/security/resources/compass.jpg. This content was automatically upgraded and should be served over HTTPS.

This test loads a secure frame with mixed content that tries to access geolocation. Access to geolocation is blocked over secure connections with mixed content.
